+++ client/src/views/all-games/components/allGamesUtils.ts
